+++ assignment2/cs231n/kaggle/DigitRecognizer/2_model_run
