+++ assignment2/cs231n/kaggle/DigitRecognizer/DigitRecognizerCNN.py
@@ -1,12 +1,13 @@
 """ url - https://www.kaggle.com/c/digit-recognizer """
 
 from time import time
-import numpy as np
+
+import matplotlib.cm as cm
 import matplotlib.pyplot as plt
 import pandas as pd
+
 from cs231n.classifiers.DigitRecognizerConvNet import *
-from cs231n.kaggle.DigitRecognizerConvNetSolver import Solver
-import matplotlib.cm as cm
+from cs231n.kaggle.DigitRecognizer.DigitRecognizerConvNetSolver import Solver
 
 plt.rcParams['figure.figsize'] = (10.0, 8.0)  # set default size of plots
 plt.rcParams['image.interpolation'] = 'nearest'
