+++ assignment2/cs231n/kaggle/DigitRecognizer/DigitRecognizerConvNetSolver.py
