+++ assignment2/cs231n/kaggle/DigitRecognizer/test.py
